fixes(code style): remove primitive unboxing

+++ src/main/java/gov/nasa/engine_sim_ua/BurnerPanel.java
@@ -152,9 +152,9 @@ public class BurnerPanel extends Panel {
             switch (Turbo.mburner) {
                 case 0: {
                     V1 = Double.valueOf(burnerLeftPanel.getDb().getText());
-                    v1 = V1.doubleValue();
+                    v1 = V1;
                     V2 = Double.valueOf(burnerLeftPanel.getTb().getText());
-                    v2 = V2.doubleValue();
+                    v2 = V2;
                     Turbo.dburner = v1 / Turbo.dconv;
                     Turbo.tburner = v2 / Turbo.tconv;
                     break;
@@ -339,17 +339,17 @@ public class BurnerPanel extends Panel {
             float fl1;
 
             V1 = Double.valueOf(getF1().getText());
-            v1 = V1.doubleValue();
+            v1 = V1;
             V2 = Double.valueOf(getF2().getText());
-            v2 = V2.doubleValue();
+            v2 = V2;
             V3 = Double.valueOf(getF3().getText());
-            v3 = V3.doubleValue();
+            v3 = V3;
             V6 = Double.valueOf(getF4().getText());
-            v6 = V6.doubleValue();
+            v6 = V6;
             V4 = Double.valueOf(getDb().getText());
-            v4 = V4.doubleValue();
+            v4 = V4;
             V5 = Double.valueOf(getTb().getText());
-            v5 = V5.doubleValue();
+            v5 = V5;
 
             // Materials
             if(Turbo.mburner == 0) {
